Generalize test for multiple backends.
